src/modules/auth/auth.service.ts
import {
  ConflictException,
  HttpException,
  Injectable,
  UnauthorizedException,
} from '@nestjs/common';
import { CreateUserDto } from '@/modules/user/dto/CreateUserDto';
import { JwtService } from '@nestjs/jwt';
import * as bcrypt from 'bcryptjs';
import { UserDocument } from '@/modules/user/schemas/User.schema';
import { UserLogin } from '@/common/types/user.interfaces';
import { Request } from 'express';
import { UserService } from '../user/user.service';

@Injectable()
export class AuthService {
  constructor(
    private userService: UserService,
    private jwtService: JwtService,
  ) {}

  async login(userDto: UserLogin) {
    const user = await this.validateEmailPass(userDto);
    return this.generateToken(user);
  }

  async register(userDto: CreateUserDto) {
    const candidate = await this.userService.getUserByEmail(userDto.email);
    if (candidate) {
      throw new ConflictException('User already exists');
    }

    const hashPassword = await bcrypt.hash(userDto.password, 5);
    const user = await this.userService.createUser({
      ...userDto,
      password: hashPassword,
    });
    return this.generateToken(user);
  }

  validateToken(token: string) {
    try {
      return this.jwtService.verify(token, {
        secret: process.env.JWT_SECRET,
      });
    } catch (e) {
      throw new UnauthorizedException({ message: 'Invalid token' });
    }
  }

  extractTokenFromHeader(request: Request): string | undefined {
    const [type, token] = request.headers.authorization?.split(' ') ?? [];
    return type === 'Bearer' ? token : undefined;
  }

  private async generateToken(user: UserDocument) {
    const payload = {
      email: user.email,
      id: user._id,
      roles: user.roles,
      username: user.username,
    };
    return {
      token: this.jwtService.sign(payload),
    };
  }

  private async validateEmailPass(userDto: UserLogin) {
    const user = await this.userService.getUserByEmail(userDto.email);
    if (!user) throw new HttpException('User not found', 404);
    const passwordEquals = await bcrypt.compare(
      userDto.password,
      user.password,
    );
    if (user && passwordEquals) {
      return user;
    }
    throw new UnauthorizedException({ message: 'Incorrect email or password' });
  }
}
